Unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1719979 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: d198c367aaa7c74c2350f620a0ede71e44fe530b

+++ src/core/org/apache/jmeter/report/processor/graph/impl/ResponseTimeDistributionGraphConsumer.java
@@ -72,7 +72,7 @@ public class ResponseTimeDistributionGraphConsumer extends
             @Override
             public Double select(Sample sample) {
                 long elapsed = sample.getElapsedTime();
-                return Double.valueOf(((double) (elapsed - elapsed % granularity)));
+                return Double.valueOf((elapsed - elapsed % granularity));
 
             }
         };
